refactor: Cleanup Profile page
