+++ frontend/src/features/profile/components/index.ts
@@ -1,7 +1,7 @@
-export { default as ProfileInfo } from './ProfileInfo';
+export { default as ProfileUser } from './ProfileUser';
 export { default as ProfileMenu } from './ProfileMenu';
 export { default as ProfileBid } from './ProfileBid';
 export { default as ProfilePoint } from './ProfilePoint';
 export { default as ProfileStats } from './ProfileStats';
-export { default as ProfileWish } from './ProfileWish';
+export { default as ProfileTransaction } from './ProfileTransaction';
 export { default as ProfileAccount } from './ProfileAccount';
