+++ frontend/src/features/profile/types/ProfileTransactionProps.ts
@@ -1,6 +1,7 @@
-export interface ProfileWishProps {
+export interface ProfileTransactionProps {
     id: number;
     name: string;
+    date: string;
     price: string;
     imageUrl: string;
 }
